+++ feature/signin/src/main/java/tht/feature/signin/StringProviderImpl.kt
@@ -22,7 +22,6 @@ class StringProviderImpl @Inject constructor(
             StringProvider.ResId.CustomerService -> R.string.message_customer_service
             StringProvider.ResId.InvalidatePhone -> R.string.message_invalidate_phone
             StringProvider.ResId.InvalidateSignupProcess -> R.string.message_invalidate_signup_process
-            StringProvider.ResId.SendAuthSuccess -> R.string.message_send_auth_success
             StringProvider.ResId.SendAuthFail -> R.string.message_send_auth_fail
             StringProvider.ResId.CheckSignupStateFail -> R.string.message_check_signup_state_fail
             StringProvider.ResId.AuthTimeout -> R.string.message_auth_timeout
