+++ feature/signin/src/main/res/layout/item_signup_custom_toast.xml
@@ -22,9 +22,9 @@
 
     <TextView
         android:layout_marginTop="30dp"
-        android:layout_marginHorizontal="60dp"
+        android:layout_marginHorizontal="20dp"
         android:gravity="center"
-        tools:text="@string/message_send_auth_success"
+        tools:text="@string/message_phone_auth_complete"
         style="@style/H2BoldTextStyle"
         android:textColor="@color/white_f9fafa"
         android:id="@+id/tv_title_custom_toast"
